Move symbolic_counting_utils to a new module
