+++ qualtran/bloqs/basic_gates/on_each.py
@@ -23,7 +23,7 @@ from qualtran import Bloq, BloqBuilder, QAny, Register, Signature, Soquet, Soque
 from qualtran.drawing import WireSymbol
 from qualtran.drawing.musical_score import TextBox
 from qualtran.resource_counting import BloqCountT, SympySymbolAllocator
-from qualtran.resource_counting.symbolic_counting_utils import SymbolicInt
+from qualtran.symbolics import SymbolicInt
 
 
 @attrs.frozen
