+++ qualtran/bloqs/basic_gates/rotation.py
@@ -23,7 +23,7 @@ from attrs import frozen
 from qualtran import bloq_example, BloqDocSpec, CompositeBloq, DecomposeTypeError
 from qualtran.cirq_interop import CirqGateAsBloqBase
 from qualtran.cirq_interop.t_complexity_protocol import TComplexity
-from qualtran.resource_counting.symbolic_counting_utils import SymbolicFloat
+from qualtran.symbolics import SymbolicFloat
 
 
 @runtime_checkable
