+++ qualtran/bloqs/qubitization_walk_operator.py
@@ -28,7 +28,7 @@ from qualtran.resource_counting.generalizers import (
     ignore_cliffords,
     ignore_split_join,
 )
-from qualtran.resource_counting.symbolic_counting_utils import SymbolicFloat
+from qualtran.symbolics import SymbolicFloat
 
 
 @attrs.frozen(cache_hash=True)
