+++ qualtran/symbolics/math_funcs.py
@@ -15,58 +15,14 @@ from typing import cast, Sized, Union
 
 import numpy as np
 import sympy
-from attrs import field, frozen, validators
-from cirq._doc import document
 
-SymbolicFloat = Union[float, sympy.Expr]
-document(SymbolicFloat, """A floating point value or a sympy expression.""")
-
-SymbolicInt = Union[int, sympy.Expr]
-document(SymbolicFloat, """A floating point value or a sympy expression.""")
-
-SymbolicComplex = Union[complex, sympy.Expr]
-document(SymbolicComplex, """A complex value or a sympy expression.""")
-
-
-@frozen
-class Shaped:
-    """Symbolic value for an object that has a shape.
-
-    A Shaped object can be used as a symbolic replacement for any object that has an attribute `shape`,
-    for example numpy NDArrays.
-    Each dimension can be either an positive integer value or a sympy expression.
-
-    This is useful to do symbolic analysis of Bloqs whose call graph only depends on the shape of the input,
-    but not on the actual values.
-    For example, T-cost of the `QROM` Bloq depends only on the iteration length (shape) and not on actual data values.
-    """
-
-    shape: tuple[SymbolicInt, ...] = field(validator=validators.instance_of(tuple))
-
-    def is_symbolic(self):
-        return True
-
-
-def is_symbolic(*args) -> bool:
-    """Returns whether the inputs contain any symbolic object.
-
-    Returns:
-        True if any argument is either a sympy object,
-        or implements the `is_symbolic` method which returns True.
-    """
-
-    if len(args) != 1:
-        return any(is_symbolic(arg) for arg in args)
-
-    (arg,) = args
-    if isinstance(arg, sympy.Basic):
-        return True
-
-    checker = getattr(arg, 'is_symbolic', None)
-    if checker is not None:
-        return checker()
-
-    return False
+from qualtran.symbolics.types import (
+    is_symbolic,
+    Shaped,
+    SymbolicComplex,
+    SymbolicFloat,
+    SymbolicInt,
+)
 
 
 def pi(*args) -> SymbolicFloat:
